ppcfd/models/ppfno/requirements.txt
# we should use this to set all the set-up files (This would be more official way to follow)
# image: registry.baidubce.com/paddlepaddle/paddle:3.0.0b1-gpu-cuda11.8-cudnn8.6-trt8.5

configmypy
# open3d==0.18.0   # use this open3d: https://github.com/PFCCLab/Open3D
h5py~=3.11.0
hydra-core
meshio

# must add 'x = x.to(torch.float32)' to /neuralop/layers/spectral_convolution.py line433
neuraloperator==0.3.0

numpy~=1.26.4
opt-einsum
pandas~=2.2.1
pynvml
pyyaml
ruamel.yaml
zarr
